Console service added to show welcome message.
