+++ src/app/app.component.ts
@@ -2,7 +2,7 @@ import { Component, inject, OnInit } from '@angular/core';
 import { HeaderComponent } from './components/layout/header/header.component';
 import { FooterComponent } from './components/layout/footer/footer.component';
 import { HomeComponent } from './components/pages/home/home.component';
-import { ResourceService } from './services/resource.service';
+import { ConsoleService } from './services/console.service';
 import { DatePipe } from '@angular/common';
 
 @Component({
@@ -18,33 +18,9 @@ import { DatePipe } from '@angular/common';
 })
 export class AppComponent implements OnInit {
 
-  private resourceService = inject(ResourceService);
+  private consoleService = inject(ConsoleService);
 
   ngOnInit() {
-    const currentYear = new Date().getFullYear();
-    const projectUrl = this.resourceService.getProjectUrl();
-    this.logConsoleWelcomeMessage(currentYear, projectUrl);
-  }
-
-  private logConsoleWelcomeMessage(year: number, projectUrl: string) {
-    setTimeout(() => {
-      console.log(`%c
-                                                                                                                         
-  ██╗    ██╗███████╗██╗      ██████╗ ██████╗ ███╗   ███╗███████╗    ███████╗███████╗██╗     ██╗      █████╗ ███████╗██╗  
-  ██║    ██║██╔════╝██║     ██╔════╝██╔═══██╗████╗ ████║██╔════╝    ██╔════╝██╔════╝██║     ██║     ██╔══██╗██╔════╝██║  
-  ██║ █╗ ██║█████╗  ██║     ██║     ██║   ██║██╔████╔██║█████╗      █████╗  █████╗  ██║     ██║     ███████║███████╗██║  
-  ██║███╗██║██╔══╝  ██║     ██║     ██║   ██║██║╚██╔╝██║██╔══╝      ██╔══╝  ██╔══╝  ██║     ██║     ██╔══██║╚════██║╚═╝  
-  ╚███╔███╔╝███████╗███████╗╚██████╗╚██████╔╝██║ ╚═╝ ██║███████╗    ██║     ███████╗███████╗███████╗██║  ██║███████║██╗  
-   ╚══╝╚══╝ ╚══════╝╚══════╝ ╚═════╝ ╚═════╝ ╚═╝     ╚═╝╚══════╝    ╚═╝     ╚══════╝╚══════╝╚══════╝╚═╝  ╚═╝╚══════╝╚═╝  
-                                                                                                                         
-  ╔═══════════════════════════════════════════════════════════════════════════════════════════════════════════════════╗  
-  ║                      The site code you're exploring might be minimized.                                           ║  
-  ║                      Check it out here: ${projectUrl}                                  ║  
-  ║                                                                                                                   ║  
-  ║                                                © ${year} Franco Dipre                                                ║  
-  ╚═══════════════════════════════════════════════════════════════════════════════════════════════════════════════════╝  
-                                                                                                                         `,
-        "font-size: 14px; color: #05f903; background-color: #191919");
-    }, 1000);
+    this.consoleService.logWelcomeMessage();
   }
 }
